+++ src_3_0_0/core/layers/js/layer2dVectorWfs.js
@@ -18,6 +18,7 @@ export default function Layer2dVectorWfs (attributes) {
 
     this.attributes = Object.assign(defaultAttributes, attributes);
     Layer2dVector.call(this, this.attributes);
+    this.prepareFeaturesFor3D(this.layer?.getSource().getFeatures());
 }
 
 Layer2dVectorWfs.prototype = Object.create(Layer2dVector.prototype);